Retain directives of redeclarations of field
In some cases, the directive of the redeclaration of a field
was being lost, and this could be fixed by swapping the order of fields.

This creates consistent behaviour, where regardless of order, all the
directives are retained. This is more predictable behaviour, because it's unclear
which directives would be retained beforehand, as it mattered the order in
which the thing was called.

TESTING: A test to verify that swapping the order of directives did not
change anything

ADDED: A simple concatenation of directive array to original in
_makeMergedFieldDefinitions. This may not be as functional as other parts of code.